# Main: deinline non performance-critical methods to improve compile speed