--- a/OgreMain/include/OgreInstanceManager.h
+++ b/OgreMain/include/OgreInstanceManager.h
@@ -279,8 +279,7 @@ namespace Ogre
         /** Returns true if settings were already created for the given material name.
             If false is returned, it means getSetting will return default settings.
         */
-        bool hasSettings( const String &materialName ) const
-        { return mBatchSettings.find( materialName ) != mBatchSettings.end(); }
+        bool hasSettings( const String &materialName ) const;
 
         /** @copydoc InstanceBatch::setStaticAndUpdate */
         void setBatchesAsStaticAndUpdate( bool bStatic );
@@ -306,15 +305,7 @@ namespace Ogre
             setCustomParameter), but there's no synchronization mechanism when
             multithreading or creating more instances, that's up to the user.
         */
-        InstanceBatchIterator getInstanceBatchIterator( const String &materialName ) const
-        {
-            InstanceBatchMap::const_iterator it = mInstanceBatches.find( materialName );
-            if(it != mInstanceBatches.end())
-                return InstanceBatchIterator( it->second.begin(), it->second.end() );
-            else
-                OGRE_EXCEPT(Exception::ERR_INVALID_STATE, "Cannot create instance batch iterator. "
-                            "Material " + materialName + " cannot be found.", "InstanceManager::getInstanceBatchIterator");
-        }
+        InstanceBatchIterator getInstanceBatchIterator( const String &materialName ) const;
     };
 } // namespace Ogre