--- a/OgreMain/src/OgreAnimable.cpp
+++ b/OgreMain/src/OgreAnimable.cpp
@@ -32,6 +32,41 @@ namespace Ogre {
     //--------------------------------------------------------------------------
     AnimableObject::AnimableDictionaryMap AnimableObject::msAnimableDictionary;
     //--------------------------------------------------------------------------
+    void AnimableObject::createAnimableDictionary(void) const
+    {
+        if (msAnimableDictionary.find(getAnimableDictionaryName()) == msAnimableDictionary.end())
+        {
+            StringVector vec;
+            initialiseAnimableDictionary(vec);
+            msAnimableDictionary[getAnimableDictionaryName()] = vec;
+        }
+    }
+
+    /// Get an updateable reference to animable value list
+    StringVector& AnimableObject::_getAnimableValueNames(void)
+    {
+        AnimableDictionaryMap::iterator i = msAnimableDictionary.find(getAnimableDictionaryName());
+        if (i != msAnimableDictionary.end())
+        {
+            return i->second;
+        }
+
+        OGRE_EXCEPT(Exception::ERR_ITEM_NOT_FOUND, "Animable value list not found for " + getAnimableDictionaryName());
+    }
+
+    const StringVector& AnimableObject::getAnimableValueNames(void) const
+    {
+        createAnimableDictionary();
+
+        AnimableDictionaryMap::iterator i = msAnimableDictionary.find(getAnimableDictionaryName());
+        if (i != msAnimableDictionary.end())
+        {
+            return i->second;
+        }
+
+        OGRE_EXCEPT(Exception::ERR_ITEM_NOT_FOUND, "Animable value list not found for " + getAnimableDictionaryName());
+    }
+
     void AnimableValue::resetToBaseValue(void)
     {
         switch(mType)