--- a/OgreMain/src/OgreAnimationState.cpp
+++ b/OgreMain/src/OgreAnimationState.cpp
@@ -409,5 +409,10 @@ namespace Ogre
         return ConstEnabledAnimationStateIterator(
             mEnabledAnimationStates.begin(), mEnabledAnimationStates.end());
     }
+
+    ControllerValueRealPtr AnimationStateControllerValue::create(AnimationState* targetAnimationState, bool addTime)
+    {
+        return std::make_shared<AnimationStateControllerValue>(targetAnimationState, addTime);
+    }
 }