--- a/OgreMain/src/OgreInstanceManager.cpp
+++ b/OgreMain/src/OgreInstanceManager.cpp
@@ -439,6 +439,10 @@ namespace Ogre
         //Return default
         return BatchSettings().setting[id];
     }
+    bool InstanceManager::hasSettings(const String& materialName) const
+    {
+        return mBatchSettings.find(materialName) != mBatchSettings.end();
+    }
     //-----------------------------------------------------------------------
     void InstanceManager::applySettingToBatches( BatchSettingId id, bool value,
                                                  const InstanceBatchVec &container )
@@ -629,4 +633,13 @@ namespace Ogre
         mesh->clearBoneAssignments();
     }
     //-----------------------------------------------------------------------
+    InstanceManager::InstanceBatchIterator InstanceManager::getInstanceBatchIterator( const String &materialName ) const
+    {
+        InstanceBatchMap::const_iterator it = mInstanceBatches.find( materialName );
+        if(it != mInstanceBatches.end())
+            return InstanceBatchIterator( it->second.begin(), it->second.end() );
+
+        OGRE_EXCEPT(Exception::ERR_INVALID_STATE, "Cannot create instance batch iterator. "
+                    "Material " + materialName + " cannot be found");
+    }
 }