… OgreMain/src/OgreStringInterface.cpp