notably StringInterface and map related methods.

# before
Compilation (524 times):
  Parsing (frontend):          931.2 s
  Codegen & opts (backend):    474.7 s

# after
Compilation (525 times):
  Parsing (frontend):          786.6 s
  Codegen & opts (backend):    450.5 s

# real-world
no significant compile time reduction on CI.